learn: seach functionalioty.auth.restrict-route

+++ base/views.py
@@ -1,7 +1,13 @@
 from django.shortcuts import render, redirect
 from django.http import HttpResponse
-from .models import Room
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
+from .models import Room, Topic
 from .forms import RoomForm
+from django.db.models import  Q
+from django.contrib.auth.models import User
+from django.contrib import  messages
+from django.contrib.auth import authenticate, login, logout
 
 
 rooms = [
@@ -14,9 +20,41 @@ rooms = [
     {'id': 7, 'name': 'Django on Steroids'},
 ]
 # Create your views here.
+def loginPage(request):
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username OR password does not exist')
+
+    context = {}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
 def home(request):
-    rooms = Room.objects.all()
-    context =  {'rooms': rooms}
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q) |
+        Q(name__icontains=q) | 
+        Q(description__icontains=q)
+        )
+    topics = Topic.objects.all()
+    room_count = rooms.count()
+    context =  {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
@@ -28,6 +66,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -38,22 +77,32 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
             form.save()
+            messages.success(request, 'Room updated successfully')
             return redirect('home')
 
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk) 
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+        
     if request.method == 'POST':
-        print(room)
         room.delete()
         return redirect('home') 
     return render(request, 'base/delete.html', {'obj': room})
